# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Library Management API",
    "description": "API Documentation for Library Management System"
  },
  "host": "localhost:7002",
  "basePath": "/api",
  "schemes": ["http"],
  "paths": {
    "/books": {
      "get": {
        "summary": "Get all books",
        "responses": {
          "200": {
            "description": "A list of books",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Book"
              }
            }
          }
        }
      }
    },
    "/members": {
      "get": {
        "summary": "Get all members",
        "responses": {
          "200": {
            "description": "A list of members",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Member"
              }
            }
          }
        }
      }
    },
    "/members/{id}/borrow": {
      "post": {
        "summary": "Borrow a book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "bookId",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "bookId": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book borrowed successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Member not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/members/{id}/return": {
      "post": {
        "summary": "Return a book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "bookId",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "bookId": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book returned successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Member not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "definitions": {
    "Book": {
      "type": "object",
      "properties": {
        "code": { "type": "string" },
        "title": { "type": "string" },
        "author": { "type": "string" },
        "stock": { "type": "number" }
      }
    },
    "Member": {
      "type": "object",
      "properties": {
        "code": { "type": "string" },
        "name": { "type": "string" },
        "borrowedBooks": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "book": { "$ref": "#/definitions/Book" },
              "borrowedDate": { "type": "string" },
              "returned": { "type": "boolean" }
            }
          }
        },
        "penaltyUntil": { "type": "string" }
      }
    }
  }
}